+++ src/egeo-demo/barrels.ts
@@ -23,10 +23,8 @@ import { StFormListDemoModule } from './st-form-list-demo/st-form-list-demo.modu
 
 import { StFullscreenLayoutDemoModule } from './st-fullscreen-layout-demo/st-fullscreen-layout-demo.module';
 import { StHeaderDemoModule } from './st-header-demo/st-header-demo.module';
-import { StHelpDemoModule } from './st-help-demo/st-help-demo.module';
 import { StHorizontalTabsDemoModule } from './st-horizontal-tabs-demo/st-horizontal-tabs-demo.module';
 import { StInfoBoxDemoModule } from './st-info-box-demo/st-info-box-demo.module';
-import { StInfoCardDemoModule } from './st-info-card-demo/st-info-card-demo.module';
 import { StInputDemoModule } from './st-input-demo/st-input-demo.module';
 import { StItemListDemoModule } from './st-item-list-demo/st-item-list-demo.module';
 import { StLabelDemoModule } from './st-label-demo/st-label-demo.module';
@@ -43,14 +41,11 @@ import { StSelectDemoModule } from './st-select-demo/select-demo.module';
 import { StSidebarDemoModule } from './st-sidebar-demo/st-sidebar-demo.module';
 import { StSpinnerDemoModule } from './st-spinner-demo/st-spinner-demo.module';
 import { StSwitchDemoModule } from './st-switch-demo/st-switch-demo.module';
-import { StTabBoxDemoModule } from './st-tab-box-demo/st-tab-box-demo.module';
 import { StTableDemoModule } from './st-table-demo/st-table-demo.module';
 import { StTagInputDemoModule } from './st-tag-input-demo/st-tag-input-demo.module';
 import { StTextareaDemoModule } from './st-textarea-demo/st-textarea-demo.module';
-import { StTipDemoModule } from './st-tip-demo/st-tip-demo.module';
 import { StToggleButtonsDemoModule } from './st-toggle-buttons-demo/st-toggle-buttons-demo.module';
 import { StTooltipDemoModule } from './st-tooltip-demo/st-tooltip-demo.module';
-import { StTreeDemoModule } from './st-tree-demo/st-tree-demo.module';
 import { StTwoListSelectionDemoModule } from './st-two-list-selection-demo/st-two-list-selection-demo.module';
 import { StVerticalTabsDemoModule } from './st-vertical-tabs-demo/st-vertical-tabs-demo.module';
 import { StWidgetDemoModule } from './st-widget-demo/st-widget-demo.module';
@@ -76,10 +71,8 @@ export const EGEO_DEMO_MODULES: any[] = [
    StFormListDemoModule,
    StFullscreenLayoutDemoModule,
    StHeaderDemoModule,
-   StHelpDemoModule,
    StHorizontalTabsDemoModule,
    StInfoBoxDemoModule,
-   StInfoCardDemoModule,
    StInputDemoModule,
    StItemListDemoModule,
    StLabelDemoModule,
@@ -96,14 +89,11 @@ export const EGEO_DEMO_MODULES: any[] = [
    StSidebarDemoModule,
    StSpinnerDemoModule,
    StSwitchDemoModule,
-   StTabBoxDemoModule,
    StTableDemoModule,
    StTagInputDemoModule,
    StTextareaDemoModule,
-   StTipDemoModule,
    StToggleButtonsDemoModule,
    StTooltipDemoModule,
-   StTreeDemoModule,
    StTwoListSelectionDemoModule,
    StVerticalTabsDemoModule,
    StZeroPageDemoModule,