+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Update icon font
 * st-table: Align cell content to the top when texts are so long
 * Remove st-info-box from demo.
+* Removed st-tree, st-tip, st-tab-box, st-info-card, st-help components
 
 
 ## 16.2.0 (April 12, 2019)